+++ apps/studio/src/constants/common.ts
@@ -5,5 +5,6 @@ export const userRoles = {
 
 export const workspaceTypes = {
   SANITY_STRUCTURE_TOOL: 'sanity-structure-tool',
+  SANITY_STRUCTURE_TOOL_SPANISH: 'sanity-structure-tool-spanish',
   TESTING: 'testing',
 } as const;