+++ docs/.vitepress/config.ts
@@ -80,9 +80,12 @@ export default defineConfig({
         items: [
           {
             text: 'Setup',
-            link: '/guide/setup',
             collapsed: true,
             items: [
+              {
+                text: 'Overview',
+                link: '/guide/setup',
+              },
               {
                 text: 'Configuration',
                 link: '/guide/setup/configuration',
@@ -90,6 +93,10 @@ export default defineConfig({
               {
                 text: 'Define List Items',
                 link: '/guide/setup/define-list-items',
+              },
+              {
+                text: 'Internationalization (i18n)',
+                link: '/guide/setup/i18n',
               },
             ],
           },
@@ -157,6 +164,10 @@ export default defineConfig({
               {
                 text: 'hideAddButton',
                 link: '/examples/hide-add-button',
+              },
+              {
+                text: 'i18n',
+                link: '/examples/i18n',
               },
               {
                 text: 'icon',